pulled tabPane into own controller. BUGGY
