+++ src/proj6BittingCerratoCohenEllmer/CodeTabs.fxml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+    File: Main.fxml
+    Authors: Caleb Bitting, Andy Xu, Jasper Loverude
+    Class: CS 361
+    Created for Project 5 on March 7
+-->
+<!-- set primary scene of size 800*600 -->
+
+<?import javafx.scene.control.TabPane?>
+<?import javafx.scene.layout.AnchorPane?>
+
+<TabPane xmlns:fx="http://javafx.com/fxml/1"
+         fx:controller="proj6BittingCerratoCohenEllmer.CodeTabController"
+         fx:id="tabPane" layoutX="14.0" layoutY="112.0"
+         prefHeight="396.0" prefWidth="800.0" tabClosingPolicy="ALL_TABS"
+         AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
+         AnchorPane.topAnchor="74.0" />
