+++ src/proj6BittingCerratoCohenEllmer/Main.fxml
@@ -13,7 +13,6 @@
 <?import javafx.scene.control.MenuBar?>
 <?import javafx.scene.control.MenuItem?>
 <?import javafx.scene.control.SeparatorMenuItem?>
-<?import javafx.scene.control.TabPane?>
 <?import javafx.scene.control.ToolBar?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import org.fxmisc.richtext.StyleClassedTextArea?>
@@ -62,11 +61,7 @@
         </Menu>
     </MenuBar>
     <!-- use a TabPane to hold the text area  -->
-    <TabPane fx:id="tabPane" layoutX="14.0" layoutY="112.0"
-             prefHeight="396.0" prefWidth="800.0" tabClosingPolicy="ALL_TABS"
-             AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0"
-             AnchorPane.topAnchor="74.0">
-    </TabPane>
+    <fx:include source="CodeTabs.fxml"/>
     <!-- use a ToolBar to hold the two buttons -->
     <ToolBar layoutY="27.0" prefHeight="49.0" prefWidth="400.0"
              AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0">
